+++ src/lib/punctuation/period.js
@@ -1,0 +1,8 @@
+
+
+
+"Sentence ending.": "Sentence ending.",
+"Sentence ending..": "Sentence ending.",
+
+/* [3] replace 2 dots at the end of the sentecne with a period */
+string = string.replace(/\.{2}/g, ".");