+++ src/locale/en-us.js
@@ -1,0 +1,10 @@
+const quotes = {
+	leftDoubleQuote : 				"“",
+	rightDoubleQuote : 				"”",
+	leftSingleQuote : 				"‘",
+	rightSingleQuote : 				"’",
+};
+
+export default {
+	quotes,
+}