+++ src/locale/locale.js
@@ -1,0 +1,15 @@
+import localeCs from "./cs";
+import localeEnUs from "./en-us";
+import localeRue from "./rue";
+import localeSk from "./sk";
+
+const locale = {
+	"cs" : localeCs,
+	"en-us" : localeEnUs,
+	"rue" : localeRue,
+	"sk" : localeSk,
+}
+
+export default {
+	locale,
+}