+++ docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for Typopo
 
+## 2.0.0 // 2017-02-
+* Complete refactoring
+* Support for (P)
+* breaking changes en > en-us
+
 ## 1.4.0 // 2017-01-15
 * “Removing empty lines” is now an optional parameter (in case you want to correct your plain-text based content or an email). By default, empty lines are removed
 * Dropped support for correcting capital letter at the beginning of the sentence. It yielded too much false positives. Feature may be re-introduced later in the future